eval (non chunked) matches (close enough)

+++ agent/eval/eval_realtime.py
@@ -1,4 +1,3 @@
-
 from agent.eval.realtime_chunking import RealtimeActionChunkingBuffer
 from agent.utils.robot_utils import get_actions, wait_for_circle
 from agent.model.policy import DiffusionPolicy
@@ -93,7 +92,7 @@ def parse_args():
     parser.add_argument('--device', type=str, default='cuda')
     parser.add_argument('--log_dir', type=str, default=None,
                         help='where to save robot log data + evaluation video (None disables logging)')
-    parser.add_argument('--control_freq', '--hz', type=int, default=10,
+    parser.add_argument('--control_freq', '--hz', type=float, default=10,
                         help='control/command frequency (Hz) for the real-time loop')
     parser.add_argument('--weight_decay', type=float, default=0.5,
                         help='recency-weighting rate (1/s) for ensembling overlapping chunks')
